AEA-1349: Move code which is not common out of the dispense common files. Fix imports. Fix tests. Additional translation fixes.
