+++ coordinator/src/services/translation/request/dispense/dispense-notification.ts
@@ -1,29 +1,31 @@
 import {fhir, hl7V3} from "@models"
 import {
+  getCodingForSystem,
   getExtensionForUrl,
+  getExtensionForUrlOrNull,
   getIdentifierValueForSystem,
   getIdentifierValueOrNullForSystem,
-  getMessageId,
   getMedicationCoding,
-  getExtensionForUrlOrNull
+  getMessageId,
+  getNumericValueAsString,
+  onlyElement
 } from "../../common"
 import {getMedicationDispenses, getMessageHeader, getPatientOrNull} from "../../common/getResourcesOfType"
 import {convertIsoDateTimeStringToHl7V3DateTime, convertMomentToHl7V3DateTime} from "../../common/dateTime"
 import pino from "pino"
 import {createAgentPersonForUnattendedAccess} from "../agent-unattended"
 import moment from "moment"
-import {DispenseNotificationSupplyHeader} from "../../../../../../models/hl7-v3"
+import {createAgentOrganisationFromReference, createPriorPrescriptionReleaseEventRef} from "./dispense-common"
+import {auditDoseToTextIfEnabled} from "../dosage"
 import {
-  createAgentOrganisation,
-  createDispenseNotificationSupplyHeaderPertinentInformation1,
-  createPrescriptionId,
-  createPrescriptionStatus,
-  createOriginalPrescriptionRef,
-  createPriorPrescriptionReleaseEventRef,
-  getOrganisationPerformer,
-  getRepeatNumberFromRepeatInfoExtension,
-  isRepeatDispensing
-} from "./dispense-common"
+  Consumable,
+  DispenseNotificationPertinentInformation2,
+  DispenseNotificationSuppliedLineItemQuantityPertinentInformation1,
+  RequestedManufacturedProduct,
+  SuppliedLineItemComponent1,
+  SupplyInstructions,
+  SupplyRequest
+} from "../../../../../../models/hl7-v3"
 
 export async function convertDispenseNotification(
   bundle: fhir.Bundle,
@@ -58,9 +60,9 @@ export async function convertDispenseNotification(
   hl7DispenseNotification.effectiveTime = convertMomentToHl7V3DateTime(moment.utc())
   hl7DispenseNotification.recordTarget = new hl7V3.RecordTargetReference(hl7Patient)
   hl7DispenseNotification.primaryInformationRecipient =
-    new hl7V3.DispenseCommonPrimaryInformationRecipient(hl7AgentOrganisation)
+    new hl7V3.DispenseNotificationPrimaryInformationRecipient(hl7AgentOrganisation)
   hl7DispenseNotification.pertinentInformation1 = hl7PertinentInformation1
-  hl7DispenseNotification.pertinentInformation2 = new hl7V3.DispenseCommonPertinentInformation2(
+  hl7DispenseNotification.pertinentInformation2 = new DispenseNotificationPertinentInformation2(
     hl7CareRecordElementCategory
   )
   if (hl7PriorMessageRef) {
@@ -98,7 +100,7 @@ async function createPertinentInformation1(
       )
     }
   )
-  const supplyHeader = new DispenseNotificationSupplyHeader(new hl7V3.GlobalIdentifier(messageId), hl7Author)
+  const supplyHeader = new hl7V3.DispenseNotificationSupplyHeader(new hl7V3.GlobalIdentifier(messageId), hl7Author)
   supplyHeader.pertinentInformation1 = hl7PertinentInformation1LineItems
   supplyHeader.pertinentInformation3 = new hl7V3.SupplyHeaderPertinentInformation3(hl7PertinentPrescriptionStatus)
   supplyHeader.pertinentInformation4 = new hl7V3.SupplyHeaderPertinentInformation4(hl7PertinentPrescriptionIdentifier)
@@ -181,4 +183,222 @@ function createPriorMessageRef(fhirHeader: fhir.MessageHeader) {
   if (replacementOf) {
     return new hl7V3.MessageRef(new hl7V3.GlobalIdentifier(replacementOf.valueIdentifier.value))
   }
+}
+
+export function createDispenseNotificationSupplyHeaderPertinentInformation1(
+  fhirMedicationDispense: fhir.MedicationDispense,
+  medicationCoding: fhir.Coding,
+  logger: pino.Logger
+): hl7V3.SupplyHeaderPertinentInformation1<hl7V3.DispenseNotificationSuppliedLineItem> {
+  const fhirPrescriptionDispenseItemNumber = getPrescriptionItemNumber(fhirMedicationDispense)
+  const fhirPrescriptionLineItemStatus = getPrescriptionLineItemStatus(fhirMedicationDispense)
+  const fhirDosageInstruction = getDosageInstruction(fhirMedicationDispense, logger)
+
+  const hl7SuppliedLineItemQuantitySnomedCode = new hl7V3.SnomedCode(
+    fhirMedicationDispense.quantity.code,
+    fhirMedicationDispense.quantity.unit
+  )
+  const hl7UnitValue = fhirMedicationDispense.quantity.value.toString()
+  const hl7Quantity = new hl7V3.QuantityInAlternativeUnits(
+    hl7UnitValue,
+    hl7UnitValue,
+    hl7SuppliedLineItemQuantitySnomedCode
+  )
+  const hl7ItemStatusCode = new hl7V3.ItemStatusCode(
+    fhirPrescriptionLineItemStatus.code,
+    fhirPrescriptionLineItemStatus.display
+  )
+  const hl7PriorOriginalItemRef = getPrescriptionItemId(fhirMedicationDispense)
+  const hl7SuppliedLineItemQuantity = createSuppliedLineItemQuantity(
+    hl7Quantity,
+    medicationCoding,
+    fhirDosageInstruction
+  )
+
+  const hl7PertinentSuppliedLineItem = new hl7V3.DispenseNotificationSuppliedLineItem(
+    new hl7V3.GlobalIdentifier(fhirPrescriptionDispenseItemNumber)
+  )
+  hl7PertinentSuppliedLineItem.consumable = new Consumable(
+    new RequestedManufacturedProduct(
+      new hl7V3.ManufacturedRequestedMaterial(
+        hl7SuppliedLineItemQuantitySnomedCode
+      )
+    )
+  )
+  hl7PertinentSuppliedLineItem.component = new hl7V3.SuppliedLineItemComponent(hl7SuppliedLineItemQuantity)
+  hl7PertinentSuppliedLineItem.component1 = new SuppliedLineItemComponent1(
+    new SupplyRequest(hl7SuppliedLineItemQuantitySnomedCode, hl7Quantity)
+  )
+  hl7PertinentSuppliedLineItem.pertinentInformation3 = new hl7V3.SuppliedLineItemPertinentInformation3(
+    new hl7V3.ItemStatus(hl7ItemStatusCode)
+  )
+  hl7PertinentSuppliedLineItem.inFulfillmentOf = new hl7V3.SuppliedLineItemInFulfillmentOf(
+    new hl7V3.OriginalPrescriptionRef(new hl7V3.GlobalIdentifier(hl7PriorOriginalItemRef))
+  )
+
+  if (isRepeatDispensing(fhirMedicationDispense)) {
+    const repeatInfo = getExtensionForUrl(
+      fhirMedicationDispense.extension,
+      "https://fhir.nhs.uk/StructureDefinition/Extension-EPS-RepeatInformation",
+      "MedicationDispense.extension"
+    ) as fhir.ExtensionExtension<fhir.IntegerExtension>
+
+    hl7PertinentSuppliedLineItem.repeatNumber = getRepeatNumberFromRepeatInfoExtension(repeatInfo)
+  }
+
+  return new hl7V3.SupplyHeaderPertinentInformation1(hl7PertinentSuppliedLineItem)
+}
+
+export function getRepeatNumberFromRepeatInfoExtension(
+  repeatInfoExtension: fhir.ExtensionExtension<fhir.IntegerExtension>
+): hl7V3.Interval<hl7V3.NumericValue> {
+  const numberOfRepeatsIssuedExtension = getExtensionForUrl(
+    repeatInfoExtension.extension,
+    "numberOfRepeatsIssued",
+    /* eslint-disable-next-line max-len */
+    "MedicationDispense.extension(\"https://fhir.nhs.uk/StructureDefinition/Extension-EPS-RepeatInformation\").extension"
+  ) as fhir.IntegerExtension
+  const numberOfRepeatsIssued = getNumericValueAsString(numberOfRepeatsIssuedExtension.valueInteger)
+  const numberOfRepeatsAllowedExtension = getExtensionForUrl(
+    repeatInfoExtension.extension,
+    "numberOfRepeatsAllowed",
+    /* eslint-disable-next-line max-len */
+    "MedicationDispense.extension(\"https://fhir.nhs.uk/StructureDefinition/Extension-EPS-RepeatInformation\").extension"
+  ) as fhir.IntegerExtension
+  const numberOfRepeatsAllowed = getNumericValueAsString(numberOfRepeatsAllowedExtension.valueInteger)
+
+  return new hl7V3.Interval<hl7V3.NumericValue>(
+    new hl7V3.NumericValue(numberOfRepeatsIssued),
+    new hl7V3.NumericValue(numberOfRepeatsAllowed)
+  )
+}
+
+export function createSuppliedLineItemQuantity(
+  hl7Quantity: hl7V3.QuantityInAlternativeUnits,
+  fhirProductCoding: fhir.Coding,
+  fhirDosageInstruction: fhir.Dosage
+): hl7V3.DispenseNotificationSuppliedLineItemQuantity {
+  const productCode = new hl7V3.SnomedCode(fhirProductCoding.code, fhirProductCoding.display)
+  const manufacturedRequestedMaterial = new hl7V3.ManufacturedRequestedMaterial(productCode)
+  const suppliedManufacturedProduct = new hl7V3.SuppliedManufacturedProduct(manufacturedRequestedMaterial)
+  const dispenseProduct = new hl7V3.DispenseProduct(suppliedManufacturedProduct)
+  const hl7SuppliedLineItemQuantity = new hl7V3.DispenseNotificationSuppliedLineItemQuantity(
+    hl7Quantity,
+    dispenseProduct
+  )
+  // eslint-disable-next-line max-len
+  hl7SuppliedLineItemQuantity.pertinentInformation1 = new DispenseNotificationSuppliedLineItemQuantityPertinentInformation1(
+    new SupplyInstructions(fhirDosageInstruction.text)
+  )
+  return hl7SuppliedLineItemQuantity
+}
+
+export function getOrganisationPerformer(fhirFirstMedicationDispense: fhir.MedicationDispense): fhir.DispensePerformer {
+  return fhirFirstMedicationDispense.performer.find(p => p.actor.type === "Organization")
+}
+
+export function getFhirGroupIdentifierExtension(
+  fhirFirstMedicationDispense: fhir.MedicationDispense
+): fhir.ExtensionExtension<fhir.Extension> {
+  const fhirAuthorizingPrescriptionExtensions =
+    fhirFirstMedicationDispense.authorizingPrescription.flatMap(e => e.extension)
+  const fhirGroupIdentifierExtension = getExtensionForUrl(
+    fhirAuthorizingPrescriptionExtensions,
+    "https://fhir.nhs.uk/StructureDefinition/Extension-DM-GroupIdentifier",
+    "MedicationDispense.authorizingPrescription") as fhir.ExtensionExtension<fhir.Extension>
+  return fhirGroupIdentifierExtension
+}
+
+export function getPrescriptionStatus(fhirFirstMedicationDispense: fhir.MedicationDispense): fhir.CodingExtension {
+  return getExtensionForUrl(
+    fhirFirstMedicationDispense.extension,
+    "https://fhir.nhs.uk/StructureDefinition/Extension-EPS-TaskBusinessStatus",
+    "MedicationDispense.extension") as fhir.CodingExtension
+}
+
+export function getPrescriptionItemId(fhirMedicationDispense: fhir.MedicationDispense): string {
+  return getIdentifierValueForSystem(
+    fhirMedicationDispense.authorizingPrescription.map(e => e.identifier),
+    "https://fhir.nhs.uk/Id/prescription-order-item-number",
+    "MedicationDispense.authorizingPrescription.identifier"
+  )
+}
+
+export function getDosageInstruction(
+  fhirMedicationDispense: fhir.MedicationDispense,
+  logger: pino.Logger
+): fhir.Dosage {
+  auditDoseToTextIfEnabled(fhirMedicationDispense.dosageInstruction, logger)
+  return onlyElement(
+    fhirMedicationDispense.dosageInstruction,
+    "MedicationDispense.dosageInstruction"
+  )
+}
+
+export function getPrescriptionItemNumber(fhirMedicationDispense: fhir.MedicationDispense): string {
+  return getIdentifierValueForSystem(
+    fhirMedicationDispense.identifier,
+    "https://fhir.nhs.uk/Id/prescription-dispense-item-number",
+    "MedicationDispense.identifier"
+  )
+}
+
+export function getPrescriptionLineItemStatus(fhirMedicationDispense: fhir.MedicationDispense): fhir.Coding {
+  return getCodingForSystem(
+    fhirMedicationDispense.type.coding,
+    "https://fhir.nhs.uk/CodeSystem/medicationdispense-type",
+    "MedicationDispense.type.coding"
+  )
+}
+
+export function createPrescriptionId(
+  fhirFirstMedicationDispense: fhir.MedicationDispense
+): hl7V3.PrescriptionId {
+  const fhirGroupIdentifierExtension = getFhirGroupIdentifierExtension(fhirFirstMedicationDispense)
+  const fhirAuthorizingPrescriptionShortFormIdExtension = getExtensionForUrl(
+    fhirGroupIdentifierExtension.extension,
+    "shortForm",
+    "MedicationDispense.authorizingPrescription.extension.valueIdentifier"
+  ) as fhir.IdentifierExtension
+  const hl7PertinentPrescriptionId = fhirAuthorizingPrescriptionShortFormIdExtension
+    .valueIdentifier
+    .value
+  return new hl7V3.PrescriptionId(hl7PertinentPrescriptionId)
+}
+
+export function createOriginalPrescriptionRef(
+  firstMedicationDispense: fhir.MedicationDispense
+): hl7V3.OriginalPrescriptionRef {
+  const fhirGroupIdentifierExtension = getFhirGroupIdentifierExtension(firstMedicationDispense)
+  const fhirAuthorizingPrescriptionShortFormIdExtension = getExtensionForUrl(
+    fhirGroupIdentifierExtension.extension,
+    "UUID",
+    "MedicationDispense.authorizingPrescription.extension.valueIdentifier"
+  ) as fhir.IdentifierExtension
+  const id = fhirAuthorizingPrescriptionShortFormIdExtension.valueIdentifier.value
+  return new hl7V3.OriginalPrescriptionRef(
+    new hl7V3.GlobalIdentifier(id)
+  )
+}
+
+export function createPrescriptionStatus(
+  fhirFirstMedicationDispense: fhir.MedicationDispense
+): hl7V3.PrescriptionStatus {
+  const prescriptionStatusExtension = getPrescriptionStatus(fhirFirstMedicationDispense)
+  const prescriptionStatusCoding = prescriptionStatusExtension.valueCoding
+  return new hl7V3.PrescriptionStatus(prescriptionStatusCoding.code, prescriptionStatusCoding.display)
+}
+
+export function isRepeatDispensing(medicationDispense: fhir.MedicationDispense): boolean {
+  return !!getExtensionForUrlOrNull(
+    medicationDispense.extension,
+    "https://fhir.nhs.uk/StructureDefinition/Extension-EPS-RepeatInformation",
+    "MedicationDispense.extension"
+  )
+}
+
+export function createAgentOrganisation(
+  organisation: fhir.DispensePerformer
+): hl7V3.AgentOrganization {
+  return createAgentOrganisationFromReference(organisation.actor)
 }
